fix things that were pointed out in https://github.com/chainer/chainerrl/pull/491
