+++ chainerrl/wrappers/__init__.py
@@ -1,7 +1,9 @@
 from chainerrl.wrappers.cast_observation import CastObservation  # NOQA
 from chainerrl.wrappers.cast_observation import CastObservationToFloat32  # NOQA
 
-from chainerrl.wrappers.continuing_time_limit import ContinuingTimeLimit, ContinuingTimeLimitMonitor  # NOQA
+from chainerrl.wrappers.continuing_time_limit import ContinuingTimeLimit  # NOQA
+
+from chainerrl.wrappers.monitor import Monitor  # NOQA
 
 from chainerrl.wrappers.randomize_action import RandomizeAction  # NOQA
