+++ examples/atari/reproduction/rainbow/train_rainbow.py
@@ -79,7 +79,7 @@ def main():
             # Randomize actions like epsilon-greedy in evaluation as well
             env = chainerrl.wrappers.RandomizeAction(env, args.eval_epsilon)
         if args.monitor:
-            env = chainerrl.wrappers.ContinuingTimeLimitMonitor(
+            env = chainerrl.wrappers.Monitor(
                 env, args.outdir,
                 mode='evaluation' if test else 'training')
         if args.render:
@@ -147,7 +147,6 @@ def main():
             eval_stats['mean'],
             eval_stats['median'],
             eval_stats['stdev']))
-        eval_env.close()
 
     else:
         experiments.train_agent_with_evaluation(
@@ -179,8 +178,6 @@ def main():
         print("The results of the best scoring network:")
         for stat in stats:
             print(str(stat) + ":" + str(stats[stat]))
-        env.close()
-        eval_env.close()
 
 
 if __name__ == '__main__':
