+++ examples/atari/train_categorical_dqn_ale.py
@@ -81,7 +81,7 @@ def main():
             # Randomize actions like epsilon-greedy in evaluation as well
             env = chainerrl.wrappers.RandomizeAction(env, args.eval_epsilon)
         if args.monitor:
-            env = chainerrl.wrappers.ContinuingTimeLimitMonitor(
+            env = chainerrl.wrappers.Monitor(
                 env, args.outdir,
                 mode='evaluation' if test else 'training')
         if args.render:
@@ -144,7 +144,6 @@ def main():
         print('n_runs: {} mean: {} median: {} stdev {}'.format(
             args.eval_n_runs, eval_stats['mean'], eval_stats['median'],
             eval_stats['stdev']))
-        eval_env.close()
     else:
         experiments.train_agent_with_evaluation(
             agent=agent, env=env, steps=args.steps,
@@ -154,8 +153,6 @@ def main():
             save_best_so_far_agent=False,
             eval_env=eval_env,
         )
-        env.close()
-        eval_env.close()
 
 
 if __name__ == '__main__':
