+++ examples/gym/train_pcl_gym.py
@@ -107,8 +107,7 @@ def main():
         # Cast observations to float32 because our model uses float32
         env = chainerrl.wrappers.CastObservationToFloat32(env)
         if args.monitor and process_idx == 0:
-            env = chainerrl.wrappers.ContinuingTimeLimitMonitor(
-                env, args.outdir)
+            env = chainerrl.wrappers.Monitor(env, args.outdir)
         if not test:
             # Scale rewards (and thus returns) to a reasonable range so that
             # training is easier
@@ -205,7 +204,6 @@ def main():
         print('n_runs: {} mean: {} median: {} stdev {}'.format(
             args.eval_n_runs, eval_stats['mean'], eval_stats['median'],
             eval_stats['stdev']))
-        env.close()
     else:
         if args.train_async:
             experiments.train_agent_async(
@@ -232,8 +230,6 @@ def main():
                 eval_n_episodes=args.eval_n_runs,
                 eval_interval=args.eval_interval,
                 train_max_episode_len=timestep_limit)
-            env.close()
-            eval_env.close()
 
 
 if __name__ == '__main__':
