+++ examples/mujoco/train_ddpg_gym.py
@@ -84,8 +84,7 @@ def main():
         # Cast observations to float32 because our model uses float32
         env = chainerrl.wrappers.CastObservationToFloat32(env)
         if args.monitor:
-            env = chainerrl.wrappers.ContinuingTimeLimitMonitor(
-                env, args.outdir)
+            env = chainerrl.wrappers.Monitor(env, args.outdir)
         if isinstance(env.action_space, spaces.Box):
             misc.env_modifiers.make_action_filtered(env, clip_action_filter)
         if not test:
@@ -168,7 +167,6 @@ def main():
         print('n_runs: {} mean: {} median: {} stdev {}'.format(
             args.eval_n_runs, eval_stats['mean'], eval_stats['median'],
             eval_stats['stdev']))
-        eval_env.close()
     else:
         experiments.train_agent_with_evaluation(
             agent=agent, env=env, steps=args.steps,
@@ -176,8 +174,6 @@ def main():
             eval_n_episodes=args.eval_n_runs, eval_interval=args.eval_interval,
             outdir=args.outdir,
             train_max_episode_len=timestep_limit)
-        env.close()
-        eval_env.close()
 
 
 if __name__ == '__main__':
