+++ examples/mujoco/train_trpo_gym.py
@@ -81,8 +81,7 @@ def main():
         # Cast observations to float32 because our model uses float32
         env = chainerrl.wrappers.CastObservationToFloat32(env)
         if args.monitor:
-            env = chainerrl.wrappers.ContinuingTimeLimitMonitor(
-                env, args.outdir)
+            env = chainerrl.wrappers.Monitor(env, args.outdir)
         if args.render:
             env = chainerrl.wrappers.Render(env)
         return env
@@ -193,7 +192,6 @@ TRPO only supports gym.spaces.Box or gym.spaces.Discrete action spaces.""")  # N
         print('n_runs: {} mean: {} median: {} stdev {}'.format(
             args.eval_n_runs, eval_stats['mean'], eval_stats['median'],
             eval_stats['stdev']))
-        env.close()
     else:
         eval_env = make_env(test=True)
         chainerrl.experiments.train_agent_with_evaluation(
@@ -207,8 +205,6 @@ TRPO only supports gym.spaces.Box or gym.spaces.Discrete action spaces.""")  # N
             eval_interval=args.eval_interval,
             train_max_episode_len=timestep_limit,
         )
-        env.close()
-        eval_env.close()
 
 
 if __name__ == '__main__':
